FO-2371: Flytt alle tester om oppfolgning i en fil
